Update project_text.js
I'm having a hard time making things transpire naturally and sequentially. I spent hours trying to make an if statement for "if positionY<0" meaning if the flowers have left the frame, but it didn't work. So now I shifted to a mousepress, and will work on a performance that incorporates the mouse click. I have been sticking to groovy colors and shapes.